LPS-164069 Remove PortletConfigurationExporterRegistry and its impl as there is only one usage of it and inline method to WidgetInstanceMapper.
